+++ include/criteria.h
@@ -59,7 +59,7 @@ bool parse_criteria(const char *string, struct mako_criteria *criteria);
 bool apply_criteria_field(struct mako_criteria *criteria, char *token);
 
 struct mako_criteria *global_criteria(struct mako_config *config);
-int apply_each_criteria(struct wl_list *criteria_list,
+ssize_t apply_each_criteria(struct wl_list *criteria_list,
 		struct mako_notification *notif);
 
 #endif
